requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pdfminer.six==20201018
python-docx==0.8.11
PyPDF2==3.0.1
